Unified user api with controller
