+++ src/main/java/com/backend/webproject/managers/UserManager.java
@@ -1,0 +1,5 @@
+package com.backend.webproject.managers;
+
+public class UserManager {
+
+}
