+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
 server.port=8081
 spring.datasource.url=jdbc:oracle:thin:@localhost:1521:XE
 spring.datasource.username=system
-spring.datasource.password=196601
+spring.datasource.password=system
 spring.datasource.driverclassname=oracle.jdbc.driver.OracleDriver
 spring.mvc.view.prefix: /WEB-INF/jsp/
 spring.mvc.view.suffix: .jsp
